perfreport-mcp/pyproject.toml
[project]
name = "perfreport-mcp"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "pypandoc>=1.13",
    "pyyaml>=6.0",
    "matplotlib>=3.8.0",
    "pandas>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services"]
include = ["perfreport.py"]

[tool.hatch.build.targets.sdist]
include = [
    "services",
    "utils",
    "templates",
    "perfreport.py",
    "pyproject.toml",
    "README.md",
    "config.yaml",
    "chart_colors.yaml"
]
